requirements.txt
pbr>=0.5.21,<1.0
Babel>=0.9.6

cliff
pyOpenSSL
python-keystoneclient>=0.4.1
six

-f http://tarballs.openstack.org/oslo.config/oslo.config-1.2.0a3.tar.gz#egg=oslo.config-1.2.0a3
oslo.config>=1.2.0a3

eventlet>=0.12.0
iso8601
netaddr
